Move catalog YAML test from rest-server to brooklyn-camp.
The test doesn't rely on REST in any way.
